Renamed secondaryKeywordAssessor to relatedKeywordAssessor.
