+++ src/worker/AnalysisWebWorker.js
@@ -30,7 +30,7 @@ import SnippetPreview from "../snippetPreview";
 import morphologyData from "../morphology/morphologyData.json";
 import Paper from "../values/Paper";
 import AssessmentResult from "../values/AssessmentResult";
-import SecondaryKeywordAssessor from "../secondaryKeywordAssessor";
+import RelatedKeywordAssessor from "../relatedKeywordAssessor";
 
 const YoastSEO = {
 	Assessor,
@@ -41,7 +41,7 @@ const YoastSEO = {
 	Pluggable,
 	Researcher,
 	SnippetPreview,
-	SecondaryKeywordAssessor,
+	RelatedKeywordAssessor,
 
 	Paper,
 	AssessmentResult,
@@ -57,7 +57,7 @@ const YoastSEO = {
 
 import CornerstoneContentAssessor from "../cornerstone/contentAssessor";
 import CornerstoneSEOAssessor from "../cornerstone/seoAssessor";
-import CornerstoneSecondaryKeywordAssessor from "../cornerstone/secondaryKeywordAssessor";
+import CornerstoneSecondaryKeywordAssessor from "../cornerstone/relatedKeywordAssessor";
 import InvalidTypeError from "../errors/invalidType";
 
 // Internal dependencies.
